Fix the six CodeQL findings Copilot's two autofixes leave open
Three int-arithmetic malloc sizes in Bio.Cluster now multiply in size_t
(alerts 3-5, the same defect class alert 16's autofix addresses in
ccealign). Bio.PDB.PICIO's segid patterns wrote [a-zA-z], a range that
also matches the six punctuation characters between Z and a; now [a-zA-Z]
(alerts 12-14). And test_Phylo's last tempfile.mktemp becomes
NamedTemporaryFile, the pattern its sibling test already uses, with the
removal moved into a finally so a failure cannot leak the file (alert 15).

+++ Bio/Cluster/cluster.c
@@ -4743,9 +4743,9 @@ yields bit-identical results on every run.
     clusterrng_init(&rng, rng_seed);
 
     if (lcelldata == 0) {
-        celldata = MALLOC(nxgrid*nygrid*ndata*sizeof(double**));
+        celldata = MALLOC((size_t)nxgrid*nygrid*ndata*sizeof(double**));
         for (i = 0; i < nxgrid; i++) {
-            celldata[i] = MALLOC(nygrid*ndata*sizeof(double*));
+            celldata[i] = MALLOC((size_t)nygrid*ndata*sizeof(double*));
             for (j = 0; j < nygrid; j++)
                 celldata[i][j] = MALLOC(ndata*sizeof(double));
         }
